owncast: Fix statedirectory issue after upgrade
The entrypoint script would ensure a symlink to
the nix store exists in the state directory.
The script would fail when trying to overwrite an
existing symlink.

Steps to trigger the issue:

* Install owncast
* Upgrade owncast
* Garbage collect the store
* Upgrade owncast

I chose not to use the -f option in `ln`, as this is
more explicit.

(cherry picked from commit ac2d6bef042e817b7c8734a2aca4ffa904996425)