acme-client: add LibreSSL 3.5 compatibility patch
LibreSSL and OpenSSL used to have a slight incompatibility up to
LibreSSL 3.4. To be compatible with both, acme-client used to have a
macro that expanded differently for LibreSSL and OpenSSL. But in
LibreSSL 3.5, LibreSSL adopted OpenSSL's interface. The compatibility
macro now actually breaks the build when building against LibreSSL 3.5.

For now it's not a problem that acme-client is incompatible with
LibreSSL 3.5, but 3.4 will stop being supported in October (1 year after
the OpenBSD 7.0 release on 2021-10-14), so make sure that we can use the
latest LibreSSL versions.

I also submitted this patch upstream [1].

[1]: https://lists.sr.ht/~graywolf/public-inbox/%3C78e2881e-05ea-ebe2-6d40-5504041c5c4d%40veniogames.com%3E