gpick: ensure GSettings schemas are available
Running gpick on a system where GSettings schemas are not provided (e.g.
XMonad with no desktop environment), any attempt to save or export color
info from gpick will cause gpick to crash:

  GLib-GIO-ERROR **: 19:12:55.095: No GSettings schemas are installed on the system

The nixpkgs manual describes how to fix such errors:
https://nixos.org/manual/nixpkgs/stable/#ssec-gnome-common-issues-no-schemas

Adding wrapGAppsHook to the nativeBuildInputs appears to be all that was
needed to fix the issue in this case.