gotags: reinit at 1.4.1
This change reverts #176044 and #178000, as well as updating the version
to the latest stable release. Unfortunately, despite the lack of updates
to gotags, it is still depended upon by vim-go, thus we must keep it
around until they (and other consumers) have removed it from the
dependency trees.

This change also adds myself as a maintainer, since I would not wish the
burden of this package's maintanence on anybody else.