cmctl: fix version encoded inside binary
PR includes the version and git revision as ld build flags to ensure
correct version is output with the `cmctl version` command.

Adds an `update.sh` script which can be used to update the version of
cmctl. Will automatically fetch the latest semvar release of cmctl, and
update hashes accordingly in `default.nix`.

Signed-off-by: joshvanl <me@joshvanl.dev>