smartmontools: always set scriptpath
configure.ac in smartmontools defines a PATH override for scripts to
use. This defaults to /usr/local/bin:/usr/bin:/bin if no path is
specified using --with-scriptpath. --with-scriptpath=no disables the
PATH override.

configureFlags doesn't include --with-scriptpath if enableMail is set to
false. This causes smartd_notify.sh to break when it tries to execute
sed.

Fix this by specifying --with-scriptpath=no when enableMail is false.