stdenv: fix usage of native check inputs when using strictDeps, fixes #161570
In the past `checkInputs` were added to `nativeBuildInputs`. This was
primarily done for convenience, since the introduction of `checkInputs`
broke testing of Python packages (which already had the concept of
`checkInputs`), and fixing all expressions directly was a hassle.

Now we do the right thing, making sure `checkInputs` is added to
`buildInputs` and also introducing `nativeCheckInputs` which is meant
for things like hooks.

Note this will require fixing all Python expressions using
`checkInputs`.

See #161570 for more details on this topic.