sigal: fix crash when building gallery
When building a gallery using `sigal build` the process would error
out with an message along the lines of

    shutil.Error: [('/nix/store/…/static/css',
                    '/…/_build/static/css',
                    "[Errno 13] Permission denied: '/…/_build/static/css'"),
                   …
                  ]

This is caused by `shutil.copytree` copying the permissions (555) of
the directories in the Nix store and then attempting to make
modifications to the copied directory. The patch makes the copy ignore
the permissions of directories.