libguestfs: fix virt-* wrappers
The virt-* scripts provided by upstream rely on a correct $0, so they
cannot be wrapped themselves. The scripts call the guestfish binary,
which is wrapped itself.

Fix this by not wrapping the scripts, but use substituteInPlace to call
the binary guestfish wrapper, which can be execed with an overridden $0
and will set all the required environment variables.