iterm2: fix on macOS 13
The fixup phase runs patch-shebangs, which modifies the shell
integration scripts included the iTerm2.app bundle. This causes iTerm2
to be identified as damaged on macOS 13, which prevents it from running.
This is due to changes in macOS 13 that checks signatures every time a
notarized app is run on macOS 13. Since iTerm2 is using upstream’s build
due to entitlements, it is notarized and must not be modified.

See https://eclecticlight.co/2022/06/17/app-security-changes-coming-in-ventura/
for more details on the changes.