yuzu-{ea,mainline}: {2841,1092} -> {2901,1131}
This also changes the update script for EA to fetch the two latest
releases, filter for tags matching /^EA-[0-9]*/ and pick the latest one.
This is required because recently the auto-updater (tag: continuous) is
the latest release, which would have been picked by the update script.

It also updates the free icons patch, which is needed due to yuzu
changing how they track licenses of files to REUSE. The comment is
updated to reflect the workflow that is needed to create the patch,
because pineapple does not include files beginning with a dot in its
source.