yuzu-{ea,mainline}: {2901,1131} -> {2907,1137}
This changes the update script to be compatible with the maintainer
script in `./maintainers/scripts/update.nix`, which will hopefully make
sure r-ryantm can update these packages automatically for us.

The script now also automatically updates the compatibility list, as
@Docteh pointed out a mirror is available on Github.

The call in `all-packages.nix` is changed as well to avoid a double
`callPackage`, which can make overriding quite a pain.

`QT_QPA_PLATFORM` is automatically set to `xcb` to avoid a crash on
wayland, thanks to @sbruder for the suggestion!