gpgme: respect the doCheck parameter
The current `gpgme` expression ignores the `doCheck` parameter because
upstream's `Makefile` runs the tests automatically as part of the
`buildPhase`.  Let's run the tests as part of the `checkPhase` iff
`doCheck` is set, like the rest of nixpkgs' packages.

In particular, with this commit, `pkgsCross.*.gpgme` will no longer
attempt to run the tests (tests are not supposed to be run when
host!=build).