dotnetCorePackages.combinePackages: refactor
This refactors the combinePackages function to properly combine the
different .NET versions as the previous version was copying already
wrapped binary which wasn't pointing to the combined directories'
binary and it also wasn't linking the /sdk-manifests directory.

A few other miscellaneous files like LICENSE.txt, nix-support dir and
ThirdPartyNotices.txt weren't being copied either which resulted in an
incomplete .NET installation in the combined directory.

The passthrough.icu property was also not being set which resulted in us
being unable to use the result of this function in buildDotnetModule.