nixos/tests/pulseaudio: fix pavucontrol detection
The test tries to detect the presence of pavucontrol's
window by looking for the tab label "Playback".
However, the OCR mechanism fails to resolve the
text, possibly due to the grey background color.

To fix this issue, we instead look for the name of the
sound device ("Dummy Output") which gets resolved by OCR.

Note: Strangely, the tab "Playback" *is* correctly
resolved when the test is run in interactive mode.
This might be due to the changed screen resolution,
but I didn't investigate further.