nixos/services.zerotierone: add localConfSourcePath option; improve docs on existing joinNetworks option
ZeroTier has optional config that it reads from /var/lib/zerotier-one/local.conf file. This change adds ability to provide that file via an option.
Also added an assert that sanity checks that this is not overwriting user-created file in there from before.

Docs update addresses comments in 61079 - the option is impure and should at least be documented as such.