berry: patch the version in the configure script
This fixes how berry internalizes its own version. By default a static 0.1.7 is
and overriden if a `.git` folder does exist. In that case `git describe` is used
and the version parsed out of it.

This PR unconditionally patches the version in the configure file as that appears
to be easier and more idiomatic than a "fake git".