libtool: fix shebang-fixing from 2.4.7 version bump
Libtool 2.4.7 changed the shebangs in `libtoolize.in` and `ltmain.sh`
from `/bin/sh` (which is a special sandbox exception) to `/usr/bin/env
sh` (which is not).

My PR (#167071) to bump the libtool version adjusted both of these
shebangs, when it should only have adjusted the one in
`libtoolize.in`.

`ltmain.sh` is not a script to be *executed* at
`libtool`-invocation-time.  It is a script which is *vendored into a
release* at `libtoolize`-invocation-time.

Thanks to @trofi for reporting this

  https://github.com/NixOS/nixpkgs/pull/167071#pullrequestreview-1079484605