wasmtime: disable tests on x86_64-darwin
Remove tests on x86_64-darwin to avoid specific false errors due to
the way Hydra runners are set up for this architecture.

On this platform, on Hydra runners we see: `SIMD support requires
SSE3, SSSE3, SSE4.1, and SSE4.2 on x86_64.` present in all failing
tests.

Thus, do not run tests on this platform to avoid false reports of this
derivation being broken, because Hydra runners are set up in a way
that this CPU features are not available.

An example of automation marking this derivation as broken because the
Hydra runs were reporting failures: 03bc5717445bbccda21d10eeecb6ded4c12d08fe.