boost/generic.nix: reference commit-hash in comment
The explanation for the conditional introduced by
61d9f201baeef4c4bb91ad8a8f5f89b747e0dfe4 is longer than a reasonable
inline comment should be.  It directed the reader to use `git blame`,
but that tends to bitrot.  Let's point the user to a specific nixpkgs
git hash.

This commit cannot be squashed into the previous commit, because a
commit cannot mention its own commit-hash (without performaing an
expensive double-sha1 preimage attack, of course).