build-support/rust: add cargoPurityFlag parameter
Cargo-driven builds within a Rust derivation must not access the
network.  Cargo offers two different flags to achieve this:

* `--frozen`, which we currently use.
* `--offline`, which is less picky about the `Cargo.lock` file.

Both of these prevent access to the network.  However, adding a
`[patch]` section to your `Cargo.toml` will cause `cargo build
--frozen` to fail, whereas `cargo build --offline` will still succeed.

In order for nixpkgs users to be able to take advantage of Cargo
`[patch]` stanzas, we need to provide a way for them to requrest
`--offline` be used instead of `--frozen`.  This commit adds a
`cargoPurityFlag` parameter which does that.

Simply changing `--frozen` to `--offline` across the board seemed too
risky to me.  The Cargo manual has the following warning:

> Beware that this may result in different dependency resolution than
> online mode. Cargo will restrict itself to crates that are
> downloaded locally, even if there might be a newer version as
> indicated in the local copy of the index.

https://doc.rust-lang.org/cargo/commands/cargo-build.html?highlight=frozen#manifest-options

https://doc.rust-lang.org/cargo/reference/overriding-dependencies.html#the-patch-section