doc/..rust.section.md: add section overview
This commit adds a brief five-bullet outline for the manual's Rust
section, which has grown extremely large.  It's easy for a reader to
be overwhelmed by it, especially because the manual's HTML formatting
doesn't include any kind of TOC outline below second-level headers.

The goal here is to get across to the user that:

- You can install rustc and cargo from nixpkgs
- Nixpkgs supports two approaches to packaging crates
- There are community-supported overlays outside of Nixpkgs