ryujinx: 1.1.223 -> 1.1.248
Its now running unit tests, and the headless frontend is packaged as
well. The latter is a bit inconsistent, on the OpenGL backend it will
always fail to create an SDL+OpenGL instance, causing a crash.

The Vulkan backend mostly works fine, although it might need a few
restarts before it decides to show graphics. Better than nothing though.