podman: add required argv0 to wrapper
In cases where podman uses conmon it passes its own path
as a command line argument in `--exit-command`. This exit
command is used for container cleanup. For it to work the path
must be the wrapper instead of the unwrapped executable.

Without this change rootless podman runs can hang for 20 seconds
before killing the container itself. The debug logs will show:

    Exceeded conmon timeout waiting for container ___ to exit