glib: move test substitutions into postPatch
Doing this in preCheck means that the tests will already have been
built before the substitutions are made, so they'll have to be rebuilt
after the substitutions.  Move the substitution earlier in the build,
to avoid this unnecessary extra work.

I've made the substitutions run even if tests aren't being run, to
avoid a repeat of the situation where a glib update breaks the tests
due to paths changing.