nixos/minecraft-server: optimize world generation inside test
Due to how complex minecraft world generation has gotten in recent
years, it now can take several minutes to complete the first generation
of a world seed, even on relatively new and powerful hardware.

We are testing if a minecraft server can run inside of a nix enviroment,
and not so much about stress testing the CI.

Test running before this change:

> (finished: waiting for TCP port 43000, in 118.49 seconds)

Test running with this change:

> (finished: waiting for TCP port 43000, in 27.88 seconds)

Choice of using `level-type` and `generate-structures` was made as they
support almost every version of minecraft. These two also make it
extremely clear what it does, compared to the more complex
`generator-settings` and all its toggles.