vcv-rack: 2.0.6 -> 2.1.2
In order to make this latest version of VCV Rack build, I had to
use VCV's fork of the RtAudio library, which has diverged from the
official RtAudio releases. I also disabled parallel builds because
they seemed to be causing non-deterministic build failures.

It is possible that the VCV team will eventually make new releases
of Rack that can again be built against the latest official release
of RtAudio; if that ever happens we can stop packaging VCV Rack
wit its own version of the library.