Intellij Idea: depend on system Go debugger delve
Given the size of the overall Intellij Idea build and after discussion
with @Mic92 we thought it was reasonable to make idea depend on the
NixOS-packaged version of the Go debugger delve, even though it is only
needed when using the Go plugin.

This is accomplished by setting `dlv.path` (see this comment on the
Jetbrains issue tracker for reference:
https://youtrack.jetbrains.com/issue/GO-11782/Run-Targets-Debugging-in-a-remote-Docker-container-on-Apple-Silicon-fails#focus=Comments-27-5280341.0-0) which instructs intellij to use a custom `dlv` binary (i.e. not the one which is bundled with Goland or downloaded by the Plugin which requires various linking fixes.

Note: this patch leaves the existing the postFixup block from the Goland
build which patches the bundled `dlv`. For Goland the patching approach
has the modest benefits of:

1. Not showing a popup balloon popup when debugging saying 'using custom dlv...`

2. Guaranteeing Goland is given access to the exact version of delve with which
   it is bundled.

1 is a fairly minor inconvenience, and in the case of 2 I think delve is reasonably
  stable at this point so is unlikely to cause a problem.

It would be possible to use the system `dlv` for Goland and the Go
plugin if consistency is preferred.

Signed-off-by: Silas Davis <silas.davis@monax.io>