nixos/peertube: add plugin support
This uses PeerTube’s plugin management scripts to install plugins from a
Nix store path instead of the npm registry.

Because they require a populated filesystem and database, it can’t do
the setup before PeerTube’s first start. Therefore, it creates two
separate systemd services. One is dependent on the files required
already existing, which then is started before PeerTube. In the case the
filesystem structure is not initialised yet, a service is started and
polls for the files to exist. After that, it continues with the plugin
installation and imperatively restarts PeerTube to ensure plugin
activation.

Co-authored-by: a-kenji <aks.kenji@protonmail.com>