glib.tests.withChecks: init
This will allow OfBorg to run the GLib tests, which it didn't do
before.  To do this correctly for the glib we're building, we need to
override it rather than taking glib as an input.
finalAttrs.finalPackage is not overrideable, so instead we need to use
overrideAttrs.  Since we can now consult finalAttrs.doCheck, there's
no longer any need for a separate doCheck option, so I've removed it
in favour of overrideAttrs.